requirements.txt
# streamlit
streamlit
langchain-openai
langchain-core
langchain-community
python-dotenv
